+++ src/main/java/com/aiqin/bms/scmp/api/bireport/domain/response/dashboard/ChannelSectorMonthSalesRespVo.java
@@ -5,13 +5,15 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @ApiModel("渠道部门月销售情况respVo")
 @Data
 public class ChannelSectorMonthSalesRespVo {
 
     @ApiModelProperty("查询部门销售额贡献率")
     @JsonProperty("sales_contribution_rate")
-    private Long salesContributionRate;
+    private BigDecimal salesContributionRate;
 
     @ApiModelProperty("查询负毛利SKU数")
     @JsonProperty("negative_margin_sku_number")
@@ -19,10 +21,10 @@ public class ChannelSectorMonthSalesRespVo {
 
     @ApiModelProperty("查询缺货率")
     @JsonProperty("out_stock_rate")
-    private Long outStockRate;
+    private BigDecimal outStockRate;
 
     @ApiModelProperty("渠道退供金额(元)")
     @JsonProperty("amount_channel_refund")
-    private Long amountChannelRefund;
+    private BigDecimal amountChannelRefund;
 
 }
